add doc

+++ src/Controller/CorbeilleController.php
@@ -2,13 +2,16 @@
 
 namespace App\Controller;
 
+use App\Security\Role;
 use App\Entity\Corbeille;
+use App\Form\Admin\CorbeilleGesLocType;
+use App\Security\CorbeilleVoter;
 use App\Form\Admin\CorbeilleType;
 use App\Manager\CorbeilleManager;
 use App\Repository\CorbeilleRepository;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 /**
  * Class CorbeilleController
@@ -39,19 +42,38 @@ class CorbeilleController extends AbstractGController
 
     /**
      * @Route("/add", name="corbeille_add", methods={"GET","POST"})
-     * @IsGranted("ROLE_GESTIONNAIRE")
      */
     public function add(Request $request)
     {
-        return $this->editAction($request, new Corbeille(), CorbeilleType::class,false);
+        $item=new Corbeille();
+        if (Role::isGestionnaire( $this->getUser()) || Role::isAdmin( $this->getUser())) {
+            $form = $this->createForm(CorbeilleType::class, $item);
+        } else {
+            $form = $this->createForm(CorbeilleGesLocType::class, $item,['extra_fields_message'=>$this->getUser()->getId()]);
+        }
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            if ($this->manager->save($item)) {
+                $this->addFlash(self::SUCCESS, self::MSG_CREATE);
+
+                return $this->redirectToRoute($this->domaine . '_edit', ['id' => $item->getId()]);
+            }
+
+            $this->addFlash(self::DANGER, self::MSG_CREATE_ERROR . $this->manager->getErrors($item));
+        }
+        return $this->render($this->domaine . '/add.html.twig', [
+            'item' => $item,
+            'form' => $form->createView(),
+        ]);        
     }
 
     /**
      * @Route("/{id}", name="corbeille_del", methods={"DELETE"})
-     * @IsGranted("ROLE_GESTIONNAIRE")
      */
     public function delete(Request $request, Corbeille $item)
     {
+        $this->denyAccessUnlessGranted(CorbeilleVoter::DELETE,$item);
         return $this->deleteAction($request, $item);
     }
 
@@ -78,10 +100,30 @@ class CorbeilleController extends AbstractGController
 
     /**
      * @Route("/{id}/edit", name="corbeille_edit", methods={"GET","POST"})
-     * @IsGranted("ROLE_GESTIONNAIRE")
      */
     public function edit(Request $request, Corbeille $item)
     {
+        $this->denyAccessUnlessGranted(CorbeilleVoter::UPDATE,$item);
+        if (Role::isGestionnaire( $this->getUser()) || Role::isAdmin( $this->getUser())) {
+            $form = $this->createForm(CorbeilleType::class, $item);
+        } else {
+            $form = $this->createForm(CorbeilleGesLocType::class, $item,['extra_fields_message'=>$this->getUser()->getId()]);
+        }
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            if ($this->manager->save($item)) {
+                $this->addFlash(self::SUCCESS, self::MSG_MODIFY);
+
+                return $this->redirectToRoute($this->domaine . '_edit', ['id' => $item->getId()]);
+            }
+
+            $this->addFlash(self::DANGER, self::MSG_MODIFY_ERROR . $this->manager->getErrors($item));
+        }
+        return $this->render($this->domaine . '/edit.html.twig', [
+            'item' => $item,
+            'form' => $form->createView(),
+        ]);          
         return $this->editAction($request, $item, CorbeilleType::class);
     }
 }
